+++ NEWS.md
@@ -2,6 +2,7 @@
 * Pull HTTP query cache out to the [httpcache](https://github.com/nealrichardson/httpcache) package and take dependency on that. Remove dependency on `digest` package (httpcache depends on it instead).
 * New vignette on [filters and exclusions](inst/doc/filters.md)
 * `combine` categories of categorical and categorical-array variables, and responses of multiple-response variables, into new derived variables
+* `mergeFork` to merge changes from a dataset fork back to its parent (or vice versa)
 * `startDate` and `endDate` attributes and setters for dataset entities (#10, #11)
 * Improved validation for "name" setting, especially for categories
 * Speed up `ncol(ds)` by removing a server request